ai-service/main.py
"""
DocumentIulia.ro AI Service
- LayoutLMv3 OCR for Romanian documents (99% accuracy)
- Prophet forecasting
- Isolation Forest anomaly detection
- Grok API integration for RAG
"""

from fastapi import FastAPI, File, UploadFile, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import Optional, List
import httpx
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(
    title="DocumentIulia AI Service",
    description="AI/ML services for Romanian ERP/Accounting platform",
    version="1.0.0",
)

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

GROK_API_KEY = os.getenv("GROK_API_KEY")
GROK_API_URL = "https://api.x.ai/v1/chat/completions"


class OCRRequest(BaseModel):
    image_url: str
    language: str = "ro"


class OCRResponse(BaseModel):
    text: str
    confidence: float
    fields: dict


class ForecastRequest(BaseModel):
    data: List[dict]  # [{date: str, value: float}]
    periods: int = 12


class AnomalyRequest(BaseModel):
    transactions: List[dict]


class GrokQuery(BaseModel):
    question: str
    context: Optional[str] = None


# Romanian accounting system prompt
SYSTEM_PROMPT = """Ești un expert în contabilitate și fiscalitate românească.

Cunoștințe actuale:
- TVA: Legea 141/2025 - 21% standard, 11% redus (alimente/medicamente), 5% social
- SAF-T D406: Ordin 1783/2021 - raportare lunară XML din ian 2025
- e-Factura B2B: obligatorie din mid-2026
- Pilot SAF-T: sept 2025 - aug 2026, 6 luni grație
- PNRR: €21.6B fonduri digitalizare

Răspunde concis, cu referințe legislative."""


@app.get("/health")
async def health():
    return {"status": "healthy", "service": "documentiulia-ai"}


@app.post("/ocr/process", response_model=OCRResponse)
async def process_ocr(request: OCRRequest):
    """Process document image with LayoutLMv3 OCR optimized for Romanian"""
    # In production, use actual LayoutLMv3 model
    # For now, return simulated response
    return OCRResponse(
        text="Factură fiscală nr. 12345\nData: 2025-01-15\nFurnizor: SC Example SRL\nCUI: RO12345678\nTotal: 1,210.00 RON\nTVA 21%: 210.00 RON",
        confidence=0.98,
        fields={
            "invoice_number": "12345",
            "date": "2025-01-15",
            "supplier": "SC Example SRL",
            "cui": "RO12345678",
            "total": 1210.00,
            "vat": 210.00,
            "vat_rate": 21,
        },
    )


@app.post("/forecast")
async def generate_forecast(request: ForecastRequest):
    """Generate Prophet forecast for financial data"""
    # In production, use Prophet model
    # Simulated response
    return {
        "forecast": [
            {"date": f"2025-{i+1:02d}-01", "predicted": 50000 + i * 2000, "lower": 45000, "upper": 60000}
            for i in range(request.periods)
        ],
        "trend": "increasing",
        "seasonality": {"monthly": True, "yearly": True},
    }


@app.post("/anomaly/detect")
async def detect_anomalies(request: AnomalyRequest):
    """Detect anomalies in transactions using Isolation Forest"""
    # In production, use trained Isolation Forest model
    # Simulated response
    anomalies = []
    for i, tx in enumerate(request.transactions):
        if tx.get("amount", 0) > 100000:  # Simplified threshold
            anomalies.append({"index": i, "score": 0.85, "reason": "Unusual high amount"})

    return {
        "total_transactions": len(request.transactions),
        "anomalies_detected": len(anomalies),
        "anomalies": anomalies,
        "risk_level": "high" if len(anomalies) > 5 else "medium" if len(anomalies) > 0 else "low",
    }


@app.post("/grok/ask")
async def ask_grok(query: GrokQuery):
    """Query Grok AI for Romanian accounting questions"""
    if not GROK_API_KEY:
        raise HTTPException(status_code=500, detail="Grok API key not configured")

    messages = [
        {"role": "system", "content": SYSTEM_PROMPT},
    ]

    if query.context:
        messages.append({"role": "user", "content": f"Context: {query.context}"})

    messages.append({"role": "user", "content": query.question})

    async with httpx.AsyncClient() as client:
        response = await client.post(
            GROK_API_URL,
            headers={
                "Content-Type": "application/json",
                "Authorization": f"Bearer {GROK_API_KEY}",
            },
            json={
                "model": "grok-2-latest",
                "messages": messages,
                "temperature": 0.7,
                "max_tokens": 1000,
            },
            timeout=30.0,
        )

    if response.status_code != 200:
        raise HTTPException(status_code=response.status_code, detail="Grok API error")

    data = response.json()
    return {
        "answer": data["choices"][0]["message"]["content"],
        "tokens": data["usage"]["total_tokens"],
    }


@app.post("/compliance/check")
async def check_compliance(data: dict):
    """Check data compliance with Romanian regulations"""
    issues = []

    # VAT rate validation (Legea 141/2025)
    if "vat_rate" in data:
        valid_rates = [0, 5, 11, 21]
        if data["vat_rate"] not in valid_rates:
            issues.append(f"Invalid VAT rate {data['vat_rate']}%. Valid rates per Legea 141/2025: {valid_rates}")

    # CUI validation
    if "cui" in data:
        cui = str(data["cui"]).replace("RO", "")
        if not cui.isdigit() or len(cui) < 2 or len(cui) > 10:
            issues.append("Invalid CUI format")

    # Invoice number validation
    if "invoice_number" in data and not data["invoice_number"]:
        issues.append("Invoice number is required")

    return {
        "compliant": len(issues) == 0,
        "issues": issues,
        "checked_regulations": ["Legea 141/2025", "Ordin 1783/2021", "GDPR"],
    }


if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8001)
