ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
transformers==4.41.2
torch==2.3.1
layoutlmv3==0.0.1
pillow==10.3.0
pdf2image==1.17.0
pytesseract==0.3.10
prophet==1.1.5
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.4
